Re-Organized/Added/Removed some internals
Tester probably does not work rn because of these changes

+++ Software/MCU/Rocket/SensorValues/Include/cartesian.h
@@ -1,5 +1,5 @@
 #pragma once
 
 struct cartesian{
-    unsigned short int x,y,z;
+    float x,y,z;
 };
